Move dream recording and intercom into their own windows
